new feat: update book

+++ src/main/java/com/example/library/dto/BookDTO.java
@@ -9,7 +9,7 @@ public class BookDTO {
     private String title;
     private Integer year;
     private String author;
-    private Long value;
+    private Double value;
     private String genre;
     private Double rating;
     private String description;
@@ -52,11 +52,11 @@ public class BookDTO {
         this.author = author;
     }
 
-    public Long getValue() {
+    public Double getValue() {
         return value;
     }
 
-    public void setValue(Long value) {
+    public void setValue(Double value) {
         this.value = value;
     }
